sensc: embed calibration tools inside `lbsensc`
 - moved calibration tools to `libsensc`
 - split calibration in two parts:
   - `calib` - procedures used to acquire calibration
   - `corr` - procedures used to perform calibration

`libsensc` compiles now to two libraries:
 - `libsensc` as usual with added correction module, not implemented yet
 - `libcalib` as `libsensc` enchanced with calibration creation tools

JIRA: PP-89
