+++ libs/libhmap/Makefile
@@ -1,0 +1,33 @@
+#
+# Makefile for calibration application
+#
+# This source code and Makefile is designed for compilation from phoenix-pilot-core
+#
+# Copyright 2022 Phoenix Systems
+#
+
+NAME := libhmap
+LOCAL_SRCS := hmap.c
+include $(static-lib.mk)
+
+# NAME := libcalibcore
+# LOCAL_SRCS := hmap.c calibcore.c magiron/magiron_common magmot/magmot_common
+# LOCAL_HEADERS := hmap.h calibcore.h magmot/magmot.h magiron/magiron.h
+# include $(static-lib.mk)
+
+
+# NAME := libcorr
+# CAL_MAGIRON := magiron/magiron_corr.c magiron/magiron_common.c
+# CAL_MAGMOT := magmot/magmot_corr.c magmot/magmot_common.c
+# LOCAL_SRCS := corr.c $(CAL_MAGIRON) $(CAL_MAGMOT)
+# LOCAL_HEADERS := corr.h 
+# DEPS := libalgeb libcalibcore
+# include $(static-lib.mk)
+
+
+# NAME := libcalib
+# CAL_MAGIRON := magiron/magiron_calib.c magiron/magiron_common.c
+# CAL_MAGMOT := magmot/magmot_calib.c magmot/magmot_common.c
+# LOCAL_SRCS := calib.c $(CAL_MAGIRON) $(CAL_MAGMOT)
+# DEPS := libalgeb libsensc libmctl libcalibcore
+# include $(static-lib.mk)
