+++ libs/libhmap/hmap.c
