+++ libs/sensc/magiron/magiron.h
